refactor: Kas Service UseCase 생성
util service 또한 usecase로 캡슐화

Resolves:
Related to:
